backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
defusedxml==0.7.1
Django==4.2
django-celery-results==2.5.1
django-cors-headers==3.14.0
django-sendfile==0.3.11
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.2.4
oauthlib==3.2.2
Pillow==9.5.0
prompt-toolkit==3.0.38
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.0
